// lib/api-client/models/MediaInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MediaInput = {
    url: string;
    type: MediaInput.type;
};
export namespace MediaInput {
    export enum type {
        IMAGE = 'image',
        VIDEO = 'video',
    }
}